requirements/base.txt
django==2.1
django-rest-framework
